% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.vcf-package.R
\docType{package}
\name{convert.vcf-package}
\alias{convert.vcf}
\alias{convert.vcf-package}
\title{convert.vcf: Conversion of VCF files into FinalistDX format}
\description{
Conversion of VCF files into FinalistDX format.
}
\author{
\strong{Maintainer}: Jiri Novotny \email{jiri.novotny@fno.cz}

}
