% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarek_vcf.R
\name{convert_vcf_df_to_finalist}
\alias{convert_vcf_df_to_finalist}
\title{Convert a VCF tibble to the FinalistDX format}
\usage{
convert_vcf_df_to_finalist(vcf_df)
}
\arguments{
\item{vcf_df}{A tibble: obtained with \code{\link[=read_vcf]{read_vcf()}}}
}
\value{
A tibble
}
\description{
Convert a VCF tibble to the FinalistDX format
}
