% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fno.R-package.R
\docType{package}
\name{fno.R-package}
\alias{fno.R}
\alias{fno.R-package}
\title{fno.R: Various functions and scripts for Faculty Hospital Ostrava}
\description{
Various functions and scripts for Faculty Hospital Ostrava.
}
\author{
\strong{Maintainer}: Jiri Novotny \email{jiri.novotny@fno.cz}

}
