% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarek_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF file into tidy tibble}
\usage{
read_vcf(file, sample_name, vcf_annotation_tool = c("vep", "common", "snpeff"))
}
\arguments{
\item{file}{A character scalar: path to file VCF file (can be also gzipped)}

\item{sample_name}{A character scalar: sample name to fill in \code{SAMPLE} column of the output tibble}

\item{vcf_annotation_tool}{A character scalar: which variant annotation tool was used:
\itemize{
\item \code{vep}: Variant Effect Predictor
\item \code{snpeff}: snpEff
\item \code{common}: a generic VCF
}}
}
\value{
A tibble, with \code{meta} attribute storing some additional VCF metadata
}
\description{
Internally, this function uses a \href{https://github.com/gorgitko/vcfR}{modified version} of the vcfR package
that is able to parse annotated VCF (by VEP or snpEff tools) into a tidy tibble.
It is necessary to select the proper annotation tool of origin (see the \code{vcf_annotation_tool} parameter)
in order to correctly parse the VCF.
}
